# go/model_cmd_output.go
/*
 * Swagger Kubechat
 *
 * Wrapper API of kubectl CLI command
 *
 * API version: 0.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package kubechat

type CmdOutput struct {
	Cmd string `json:"cmd"`

	Code int `json:"code"`

	Stdout []string `json:"stdout,omitempty"`

	Stderr []string `json:"stderr,omitempty"`

	ResponseTime string `json:"responseTime,omitempty"`
}