# go/model_error.go
/*
 * Swagger Kubechat
 *
 * Wrapper API of kubectl CLI command
 *
 * API version: 0.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package kubechat

type Error struct {

	Code int32 `json:"code"`

	Message string `json:"message"`
}