# go/model_kubectl.go
/*
 * Swagger Kubechat
 *
 * Wrapper API of kubectl CLI command
 *
 * API version: 0.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package kubechat

type Kubectl struct {
	Version string `json:"version"`

	AllowPattern string `json:"allowPattern"`

	Cmd string `json:"cmd"`
}